requirements.txt
streamlit
ultralytics
torch
torchvision
openai-clip
chromadb
Pillow
